Create pig_game.java
This program plays the game pig, a fairly simple gambling game that can be looked up on Wikipedia. It's quite simple and has two players: computer and user. The user has two options on his/her turn to either hold or roll. The computer has the same choice but obviously it is automated with a really lame algorithm that doesn't hold until it gets it gets 20 points during it's turn.